# reducing allocation count