@mschoch 2015-10-16
Sorry this one took so long.  The changes seem like obvious improvements.  Frustratingly they are hard to quantify in some of the longer-running benchmark programs.  BoltDB appears slower in practice (as hinted at in the micro-benchmarks above).

I'm going to merge it anyway, I believe this is an improvement, though small, and its just difficult to observe through the natural variability of the longer-running tests.

Thanks for the contribution.